requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dateutils==0.6.12
Django==2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djongo==1.3.4
idna==2.10
jugaad-data==0.22
lxml==4.6.3
multitasking==0.0.9
nsetools==1.0.11
numpy==1.20.2
pandas==1.2.4
PyJWT==1.7.1
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.2.4
typing-extensions==3.7.4.3
urllib3==1.26.4
yfinance==0.1.59
